test(BA-5821): consolidate project component tests under tests/component/project
- Move user_search_by_project tests into tests/component/project/ alongside
  the existing unassign_users tests
- Update conftest fixtures to add target_project_fixture (a project where
  regular_user is NOT pre-bound) so project.assign_users actually exercises
  the full UserRoleRow assignment path
- Switch test_project_unassign_users assertions from association_groups_users
  to association_scopes_entities to match the BA-5821 contract
