--- tests/component/user_membership/BUILD
@@ -1,5 +1,0 @@
-python_test_utils()
-
-python_tests(
-    name="tests",
-)
